Major style upgrades and score keeping.
First, CSS was utilized on the pages with a stronger utilization of the
template heirarchy. It now looks blue like a Jeopardy game, the links
for questions make a nice grid, and the links are generally block shaped
to help with clicking.

Second, session management was added so that quizzes could be started
and restarted. The server now tracks which questions a participant has
answered and the current score for correct answers. Further, answered
questions are no longer displayed on the quiz page.