Settings clean up and logging.
Changed environment variable from HOSTNAME to SERVER_HOST_NAME. There
was a name clash with VSCode's devcontainer's HOSTNAME environment
variable.

The Django project now automatically populates missing DEBUG, SECRET,
and SERVER_HOST_NAME with default parameters. This is useful for
static file generation and testing which do not need valid settings for
these variables.

The devcontainer now includes Git and Docker. However, all this allows is
building the containers. Since the containers need to mount a volume in
the local directory, we can't run in the devcontainer. But it's a step in the
right direction!

Tasks! Now there are VSCode tasks to simplify setup and running tests
and the development server.